pyproject.toml
[project]
name = "synthetic-data-service"
version = "0.1.0"
description = "FastAPI service for generating synthetic data using AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "pandas==2.1.3",
    "faker==19.12.0",
    "openai==1.3.7",
    "python-dotenv==1.0.0",
]
